feat: 회원가입 폼 관련 타입 및 초기값 정의를 위한 파일 정리
