--- src/type/joinMembership.ts
@@ -1,8 +1,0 @@
-export interface SignupValues {
-  name: string;
-  phone: string;
-  email: string;
-  emailCode: string;
-  password: string;
-  passwordCheck: string;
-}
