--- src/type/signupInitialForm.ts
@@ -1,10 +1,0 @@
-import type { JoinMembershipFormValues } from './joinMembership';
-
-export const signupInitialForm: JoinMembershipFormValues = {
-  name: '',
-  phone: '',
-  email: '',
-  emailCode: '',
-  password: '',
-  passwordCheck: ''
-};
